Add files via upload
Your client is a large MNC and they have 9 broad verticals across the organization. One of the problems your client is facing is around identifying the right people for promotion (only for the manager position and below) and prepare them in time. Currently the process, they are following is:
1. They first identify a set of employees based on recommendations/ past performance.
2. Selected employees go through a separate training and evaluation program for each vertical. These programs are based on the required skill of each vertical
3. At the end of the program, based on various factors such as training performance, KPI completion (only employees with KPIs completed greater than 60% are considered), etc., the employee gets a promotion.
For the above-mentioned process, the final promotions are only announced after the evaluation and this leads to a delay in transition to their new roles.
They have provided multiple attributes around employees' past and current performance along with demographics. Now, The task is to predict whether a potential promotee at a checkpoint in the test set will be promoted or not after the evaluation process.